requirements.txt
appdirs==1.4.4
astroid==2.5.1
attrs==20.3.0
Cerberus==1.3.2
cfgv==3.2.0
distlib==0.3.1
filelock==3.0.12
flake8==3.9.0
identify==2.2.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
nodeenv==1.5.0
packaging==20.9
paho-mqtt==1.5.1
pluggy==0.13.1
pre-commit==2.11.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.0
pylint==2.7.2
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
virtualenv==20.4.3
wrapt==1.12.1
